requirements.txt
cached_property>=1.3.0
django-localflavor>=1.4.1
django>=1.11
djangorestframework>=3.7
